fnx/templates-mcp/templates/python/BlobInputBinding/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
azurefunctions-extensions-bindings-blob